+++ model/invoice.go
@@ -1,7 +1,7 @@
 package model
 
 type Invoice struct {
-	OBUID    int     `json:"obuid"`
+	OBUID    int64     `json:"obuid"`
 	Amount   float64 `json:"amount"`
 	DateTime string  `json:"dateTime"`
 	distance float64
